data_management/monitor/requirements.txt
# DataBao 监控服务依赖包

# Web框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP客户端
aiohttp>=3.9.0
httpx>=0.25.0

# 数据处理
pydantic>=2.5.0
python-multipart>=0.0.6

# YAML配置
pyyaml>=6.0

# 系统监控
psutil>=5.9.0

# 日志
structlog>=23.0.0

# 时间处理
python-dateutil>=2.8.0

# 类型检查
typing-extensions>=4.8.0